refactor(nc-gui): add column utils file
